+++ RELEASE_NOTES.md
@@ -16,6 +16,7 @@ was added as a dependency.
 
 ## Individual updates
 
+- [#586](https://github.com/IAMconsortium/pyam/pull/586) Improve error reporting for non-floatable values in value columns
 - [#585](https://github.com/IAMconsortium/pyam/pull/585) Include revisions to the ORE manuscript source code following acceptance/publication 
 - [#583](https://github.com/IAMconsortium/pyam/pull/583) Add profiler module for performance benchmarking
 - [#579](https://github.com/IAMconsortium/pyam/pull/579) Increase performance of IamDataFrame initialization